refactor(blog): rename app
Rename Posts app into Blog for clarity
