+++ .github/workflows/lint.yml
@@ -17,7 +17,7 @@ jobs:
         run: pip install -r requirements.txt
 
       - name: Run Ruff
-        run: ruff check --output-format=github furai/ posts/
+        run: ruff check --output-format=github furai/ blog/
 
       - name: Run djLint
-        run: djlint furai/templates posts/templates --profile=django --ignore=H031
+        run: djlint furai/templates blog/templates --profile=django --ignore=H031
