+++ blog/urls.py
@@ -3,6 +3,6 @@ from django.urls import URLPattern, URLResolver, path
 from .views import index, post
 
 urlpatterns: list[URLResolver | URLPattern] = [
-    path("", index, name="posts"),
+    path("", index, name="blog"),
     path("<str:slug>", post, name="post"),
 ]
