+++ furai/settings.py
@@ -45,7 +45,7 @@ INSTALLED_APPS = [
     "django.contrib.sessions",
     "django.contrib.messages",
     "django.contrib.staticfiles",
-    "posts",
+    "blog",
 ]
 
 MIDDLEWARE = [
