+++ furai/templates/furai/header.html
@@ -3,7 +3,7 @@
   <nav class="text-white">
     <ul class="text-sm flex gap-3">
       <a class="underline" href="{% url 'home' %}">Home</a>
-      <a class="underline" href="{% url 'posts' %}">Blog</a>
+      <a class="underline" href="{% url 'blog' %}">Blog</a>
       <a class="underline">Gallery</a>
       <a class="underline">Contact</a>
     </ul>
